Fix code in .js file causing wrong score.
Previously determined the verdict of a round by extracting substring
'win' or 'lose' from result string and using them as if-conditions.
Realized afterwards that only a three-letter substring was extracted
each time, and so the extracted substring 'los' was slipping past
conditions checking for 'lose'.

Changed the code so that only the first letter of the verdict is
extracted, ie 'w' for win and 'l' for lose. Letter is then checked
against updated if-conditions.