Refactor js code
game() function was removed as it was obsolete with regards to UI.
Code was refactored to take advantage of DOM manipulation. Each time
the rock, paper, or scissors buttons is clicked the playRounds() function
triggers with the button choice serving as the player's selection.
When the round is finished, the updateScore() function updates the score
displayed on the scoreboard.
If either the user or the computer reaches 5 points the updateScore()
function declares the winner on the scoreboard and sets the boolean
noWinnerYet to false, which then skips the code block in the playRounds()
function, thereby ending the game.